Rename controllers module to PID, minor refactor
